+++ app/components/canvas/types.ts
@@ -1,61 +1,22 @@
-import { BaseEditor } from "slate";
-import { ReactEditor } from "slate-react";
+import type {
+	CanvasEditor,
+	CanvasElement,
+	CanvasText,
+} from "@/lib/canvas/types";
 
-export type ResultKind = "image" | "video" | "audio";
-
-export type CanvasElementType =
-	| "narration"
-	| "character"
-	| "image"
-	| "clip"
-	| "sound"
-	| "music";
-
-export const CANVAS_ELEMENT_TYPES = new Set<CanvasElementType>([
-	"narration",
-	"character",
-	"image",
-	"clip",
-	"sound",
-	"music",
-]);
-
-export const SCENE_TYPE = "scene" as const;
-
-export const FOREGROUND_TYPES: ReadonlySet<CanvasElementType> = new Set([
-	"image",
-	"clip",
-]);
-
-export type CanvasEditor = BaseEditor & ReactEditor & { id?: string };
-
-export type CanvasContentElement = {
-	id: string;
-	type: CanvasElementType;
-	customAttributes?: Record<string, string>;
-	children: CanvasText[];
-};
-
-export type SceneElement = {
-	id: string;
-	type: typeof SCENE_TYPE;
-	children: CanvasContentElement[];
-};
-
-export type CanvasElement = SceneElement | CanvasContentElement;
-
-export type CanvasText = {
-	id: string;
-	type: CanvasElementType;
-	text: string;
-};
-
-export type ParsedElement = {
-	id: string;
-	type: string;
-	customAttributes?: Record<string, string>;
-	children: { id: string; type: string; text: string }[];
-};
+export {
+	CANVAS_ELEMENT_TYPES,
+	FOREGROUND_TYPES,
+	SCENE_TYPE,
+	type CanvasContentElement,
+	type CanvasEditor,
+	type CanvasElement,
+	type CanvasElementType,
+	type CanvasText,
+	type ParsedElement,
+	type ResultKind,
+	type SceneElement,
+} from "@/lib/canvas/types";
 
 export {
 	METADATA_TAG_TYPES,
